timeseries: Have one <option> per language
The content of an <option> isn't really HTML, and the translation
machinery hiding <span>'s doesn't have an affect.

This works for the most part, but isn't ideal. The selected element won't
switch on language change, and not all browsers will hide other languages.

Multiple select boses may be a better solution.