dimension_timeseries: Remove TODO
I think the point here was that one should preserve the start month
when flipping between quarterly/monthly (e.g.). The int field was a
select at one point.

However, this isn't particularly irritating I don't think. The bigger
problem is that data doesn't stay in it's month when switching delta,
which is a bigger annoyance than this.